requirements.txt
pandas
numpy
openai-whisper
moviepy